debugged and validated chat ... added random client ID feature
